Remove cnnect from ClickToSend since we do not pass any params
